RUM-12980: Centralize safe read logic on the `mach_sampling_profiler`
