+++ DatadogProfiling/Mach/include/module.modulemap
@@ -19,6 +19,7 @@ module DatadogMachProfiler {
 
   explicit module Testing {
       header "ctor_profiler_testing.h"
+      header "safe_read_testing.h"
       export *
   }
 }
