--- DatadogProfiling/Tests/SafeReads/DatadogProfilingTests-Bridging-Header.h
@@ -1,7 +1,0 @@
-/*
- * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
- * This product includes software developed at Datadog (https://www.datadoghq.com/).
- * Copyright 2019-Present Datadog, Inc.
- */
-
-#import "SafeRead.h"
