+++ Package.swift
@@ -238,9 +238,6 @@ let package = Package(
                 .target(name: "TestUtilities"),
             ],
             path: "DatadogProfiling/Tests",
-            exclude: [
-                "SafeReads"
-            ],
             swiftSettings: [.interoperabilityMode(.Cxx)] + internalSwiftSettings
         ),
